Merge

+++ crce-repository-impl/src/main/java/cz/zcu/kiv/crce/repository/internal/FilebasedStoreImpl.java
@@ -2,6 +2,7 @@ package cz.zcu.kiv.crce.repository.internal;
 
 import cz.zcu.kiv.crce.metadata.Repository;
 import cz.zcu.kiv.crce.metadata.Resource;
+import cz.zcu.kiv.crce.metadata.ResourceCreator;
 import cz.zcu.kiv.crce.metadata.WritableRepository;
 import cz.zcu.kiv.crce.metadata.dao.RepositoryDAO;
 import cz.zcu.kiv.crce.metadata.dao.ResourceDAO;
@@ -14,7 +15,6 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.util.Dictionary;
-import java.util.Hashtable;
 import java.util.List;
 import java.util.Properties;
 import org.codehaus.plexus.util.FileUtils;
@@ -25,8 +25,8 @@ import org.osgi.service.event.EventHandler;
 import org.osgi.service.log.LogService;
 
 /**
- *
- * @author Jiri Kucera (kalwi@students.zcu.cz, kalwi@kalwi.eu)
+ * Filebased implementation of <code>Store</code>.
+ * @author Jiri Kucera (kalwi@students.zcu.cz, jiri.kucera@kalwi.eu)
  */
 public class FilebasedStoreImpl implements Store, EventHandler {
     
@@ -50,27 +50,29 @@ public class FilebasedStoreImpl implements Store, EventHandler {
         }
     }
 
+    /*
+     * Called by dependency manager
+     */
+    @SuppressWarnings({"UseOfObsoleteCollectionType", "unchecked"})
     void init() {
-        Dictionary props = new Hashtable();
+        Dictionary props = new java.util.Hashtable();
         props.put(EventConstants.EVENT_TOPIC, PluginManager.class.getName().replace(".", "/") + "/*");
-        props.put(EventConstants.EVENT_FILTER, "(types=*" + ResourceDAO.class.getName() + "*)");
+        props.put(EventConstants.EVENT_FILTER, "(" + PluginManager.PROPERTY_PLUGIN_TYPES + "=*" + ResourceDAO.class.getName() + "*)");
         m_context.registerService(EventHandler.class.getName(), this, props);
-        loadRepository();
     }
-
-    
     
     @Override
-    public void handleEvent(Event event) {
-        
+    public void handleEvent(final Event event) {
+        final Object lock = this;
         new Thread(new Runnable() {
 
             @Override
             public void run() {
-                loadRepository();
+                synchronized (lock) {
+                    m_repository = null;
+                }
             }
         }).start();
-        
     }
     
     private synchronized void loadRepository() {
@@ -80,18 +82,22 @@ public class FilebasedStoreImpl implements Store, EventHandler {
             m_repository = rd.getRepository(m_baseDir.toURI());
         } catch (IOException ex) {
             m_log.log(LogService.LOG_ERROR, "Could not get repository for URI: " + m_baseDir.toURI(), ex);
+            m_repository = m_pluginManager.getPlugin(ResourceCreator.class).createRepository(m_baseDir.toURI());
         }
     }
     
     public synchronized Resource move(Resource resource) throws IOException, RevokedArtifactException {
         Resource tmp = m_pluginManager.getPlugin(ActionHandler.class).beforePutToStore(resource, this);
         if (resource == null) {
-            return resource;
+            return null;
         }
         if (tmp == null) {
-            m_log.log(LogService.LOG_ERROR, "ActionHandler onPutToStore returned null resource, using original");
+            m_log.log(LogService.LOG_ERROR, "ActionHandler onPutToStore() returned null resource, using original");
         } else {
             resource = tmp;
+        }
+        if (m_repository == null) {
+            loadRepository();
         }
         if (m_repository.contains(resource)) {
             throw new RevokedArtifactException("Resource with the same symbolic name and version already exists in Store: " + resource.getId());
@@ -114,37 +120,44 @@ public class FilebasedStoreImpl implements Store, EventHandler {
         } else {
             resource = tmp;
         }
+        if (m_repository == null) {
+            loadRepository();
+        }
         if (m_repository.contains(resource)) {
             throw new RevokedArtifactException("Resource with the same symbolic name and version already exists in Store: " + resource.getId());
         }
         if ("file".equals(resource.getUri().getScheme())) {
-            return putFile(resource, false);
+            return m_pluginManager.getPlugin(ActionHandler.class).afterPutToStore(putFile(resource, false), this);
         } else {
-            return putAnother(resource, false);
+            return m_pluginManager.getPlugin(ActionHandler.class).afterPutToStore(putAnother(resource, false), this);
         }
     }
     
     private Resource putFile(Resource resource, boolean move) throws IOException, RevokedArtifactException {
         File sourceFile = new File(resource.getUri());
+        if (!sourceFile.exists()) {
+            throw new RevokedArtifactException("File to be put tu store does not exist: " + sourceFile.getPath());
+        }
+        if (m_repository == null) {
+            loadRepository();
+        }
         File targetFile = File.createTempFile("res", "", m_baseDir);
         
         ResourceDAO resourceDao = m_pluginManager.getPlugin(ResourceDAO.class);
         
-        Resource out = resourceDao.moveResource(resource, targetFile.toURI());
         if (move) {
             FileUtils.rename(sourceFile, targetFile);
         } else {
             FileUtils.copyFile(sourceFile, targetFile);
         }
-
+        Resource out = resourceDao.moveResource(resource, targetFile.toURI());
         if (!m_repository.addResource(out)) {
             m_log.log(LogService.LOG_WARNING, "Resource with the same symbolic name and version already exists in Store, but it has not been expected: " + targetFile.getPath());
             if (!targetFile.delete()) {
                 throw new IOException("Can not delete file of revoked artifact: " + targetFile.getPath());
             }
-            return out;
+            return null;
         }
-        
         resourceDao.save(out);
         
         m_pluginManager.getPlugin(RepositoryDAO.class).saveRepository(m_repository);
@@ -161,9 +174,11 @@ public class FilebasedStoreImpl implements Store, EventHandler {
         resource = m_pluginManager.getPlugin(ActionHandler.class).beforeDeleteFromStore(resource, this);
         
         if (!isInStore(resource)) {
-            if (m_repository.contains(resource)) {
+            if (m_repository != null && m_repository.contains(resource)) {
                 m_log.log(LogService.LOG_WARNING, "Removing resource is not in store but it is in internal repository: " + resource.getId());
+                m_repository.removeResource(resource);
             }
+            m_pluginManager.getPlugin(ActionHandler.class).afterDeleteFromStore(resource, this);
             return false;
         }
         
@@ -180,17 +195,23 @@ public class FilebasedStoreImpl implements Store, EventHandler {
             // once the artifact file was removed, the resource has to be removed
             // from the repository even in case of exception on removing metadata
             // to keep consistency of repository with stored artifact files
+            if (m_repository == null) {
+                loadRepository();
+            }
             if (!m_repository.removeResource(resource)) {
-                m_log.log(LogService.LOG_WARNING, "Buffer's internal repository does not contain removing resource: " + resource.getId());
+                m_log.log(LogService.LOG_WARNING, "Store's internal repository does not contain removing resource: " + resource.getId());
             }
             m_pluginManager.getPlugin(RepositoryDAO.class).saveRepository(m_repository);
         }
-        
+        m_pluginManager.getPlugin(ActionHandler.class).afterDeleteFromStore(resource, this);
         return true;
     }
 
     @Override
-    public synchronized Repository getRepository() {
+    public Repository getRepository() {
+        if (m_repository == null) {
+            loadRepository();
+        }
         return m_repository;
     }
